# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NusaAlert API",
        "description": "API dokumentasi untuk NusaAlert.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/v1/lokasi": {
            "get": {
                "tags": [
                    "Lokasi"
                ],
                "summary": "Daftar semua lokasi user",
                "operationId": "e715704fbc6097feee1cdae89b45b76d",
                "responses": {
                    "200": {
                        "description": "Daftar lokasi berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lokasi"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lokasi"
                ],
                "summary": "Tambah lokasi baru",
                "operationId": "ced68c80703dfac4fa47bb21e2a9ec2c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_lokasi",
                                    "latitude",
                                    "longitude",
                                    "radius_km"
                                ],
                                "properties": {
                                    "nama_lokasi": {
                                        "type": "string",
                                        "example": "Rumah Utama"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -6.2088
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 106.8456
                                    },
                                    "radius_km": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lokasi berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/lokasi/{id}": {
            "get": {
                "tags": [
                    "Lokasi"
                ],
                "summary": "Detail lokasi",
                "operationId": "548862b5b9f03806ac8ababe8ef8318e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail lokasi"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Lokasi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Lokasi"
                ],
                "summary": "Update lokasi",
                "operationId": "859d6b6b609e1913c6b2cbd4c42eca77",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_lokasi": {
                                        "type": "string"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "radius_km": {
                                        "type": "integer"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lokasi berhasil diperbarui"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lokasi"
                ],
                "summary": "Hapus lokasi",
                "operationId": "72d16fc343dabe73876a8cb26e377460",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lokasi berhasil dihapus"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/lokasi/{id}/toggle": {
            "patch": {
                "tags": [
                    "Lokasi"
                ],
                "summary": "Toggle status aktif lokasi",
                "operationId": "bfb974f4fcc3b4dab060fab4fe49a49e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status lokasi berhasil diubah"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Lokasi": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_lokasi": {
                        "type": "string",
                        "example": "Rumah Utama"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": -6.2088
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 106.8456
                    },
                    "radius_km": {
                        "type": "integer",
                        "example": 50
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiKey": {
                "type": "apiKey",
                "name": "X-API-Key",
                "in": "header"
            },
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
...
}